Add -public-max and -private-max flags
Add two new flags for more fine grained control over the parameter count
in a project. In addition to the -max flag, there are now the
-public-max and -private-max flags. as their name suggests, they control
the maximum number of parameters for only public/private (resp.)
functions.

The behaviour becomes somewhat complex in some use cases. For example,
if you use only -public-max or -private-max  then the other one is
ignored. So, using *only*  -public-max will result in the analysis
(effectively*) ignoring private functions. Also, when combining -max
with either -public-max or -private-max one might get a scenario where
the latter is ignored if it is larger than -max.

--
*: Technically speaking private functions with over math.MaxInt32
parameters will be reported.