Add -verbose flag
Add a -verbose flag which can be used to enable debugging logging, i.e.
additional logging that can be useful to understand the behaviour of the
program.

This is implemented by providing the analyze function with a logger,
which is set to a logger that writes to an ignored buffer when the
-verbose flag is not used, and the existing logger otherwise.