Add -excludes flag
Add a -excludes flag which can be used to exclude certain files. This
flag accepts a comma-separated list of values, each of which is a unique
pattern based on which files will be excluded.

In addition, the default number of allowed parameters has been updated
from 2 to 3, matching the (now) maximum in this project.

Lastly, there is some general refactoring of main to reduce how often
flags are read and interpreted. Instead of doing this for every provided
path in the (now removed) `analyze` function, this now happens in the
`run` function (which now calls `runWith` with baseOptions). In
addition to improving speed, this ensures that warning messages about
the flags are outputted only once.