Add one-letter aliases for common flags
Add various one-letter aliases for most/the common flags. The only flags
without an alias are -public-max and -private-max, mainly because I
couldn't think of appropriate one-letter aliases.

The usage message has been updated accordingly, and so is the CHANGELOG.

The utility function `min` required a minor update to allow for finding
the minimum value out of n values (rather than 2). So it, and its tests,
have been updated.