Convert goparamcount into gofunchecks
Convert the repository for goparamcount into a repository for
gofunchecks. gofunchecks (Go Function Checks) is intended to be a
collection of static analysis tools for Go Function, goparamcount being
one of them.

Part of the common functionality has already been extracted into the
internal/ directory. A new project level README has been added, linking
to the goparamcount README which is now in cmd/goparamcount/