Add goreturncount tool skeleton
Add a new gofunchecks tool called "goretuncount", which is similar to
goparamcount but will instead report function with too many return
values. This tool's source code can be found in the cmd/goreturncount/
directory, with its own README and CHANGELOG. The project-level README
has been updated to include goreturncount as well.

Furthermore, the testdata/ directory now has a test file for both
goparamcount and goreturncount.

NOTE: though most of the functionality of goreturncount is already
here, the core logic of checking whether a function violates the limit
or not, is not yet included in this commit. However, the test suite for
goreturncount does already exists. Therefore, when you run
`go test ./...` in this commit that will fail.